# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DreamCraft API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration",
                "operationId": "fce3d994da4149576a27585ea6d521ed",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "nickname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "first_name": "User first name",
                                    "last_name": "User last name",
                                    "nickname": "User nickname",
                                    "email": "useraddress@gmail.com",
                                    "password": "password123",
                                    "password_confirmation": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "6|6OVYnQz9AqmUmlI1AJRA5zN69WCfi9TRNh5Ud70cd81455a1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request | Duplicate email | Duplicate nickname",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Duplicate email": {
                                        "summary": "Duplicate email",
                                        "value": {
                                            "success": false,
                                            "message": "User with email already exists"
                                        }
                                    },
                                    "Duplicate nickname": {
                                        "summary": "Duplicate nickname",
                                        "value": {
                                            "success": false,
                                            "message": "User with this nickname already exist"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "The password field confirmation does not match.",
                                            "errors": {
                                                "password": "The password field confirmation does not match."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "message": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "3a335d74fc0a1eed501c656b719cc5e2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email | nickname": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "useraddress@gmail.com",
                                    "password": "password123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "6|6OVYnQz9AqmUmlI1AJRA5zN69WCfi9TRNh5Ud70cd81455a1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request | Non correct password",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Wrong email": {
                                        "summary": "Wrong email",
                                        "value": {
                                            "success": false,
                                            "message": "User doesn`t exist"
                                        }
                                    },
                                    "Wrong password": {
                                        "summary": "Wrong password",
                                        "value": {
                                            "success": false,
                                            "message": "Wrong password"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "The email field is required.",
                                            "errors": {
                                                "email": "The email field is required."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "message": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "8dc5663920cc4626176f27f2b27257ca",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "You are logged out"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "Unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "message": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}